+++ console.sh
@@ -140,7 +140,7 @@ ensure_at_least_one_model() {
   if [[ -x "$HFACE_CLI" ]]; then
     "$HFACE_CLI" download \
       TheBloke/TinyLlama-1.1B-Chat-v1.0-GGUF \
-      TinyLlama-1.1B-Chat-v1.0.Q4_K_M.gguf \
+      tinyllama-1.1b-chat-v1.0.Q4_K_M.gguf \
       --local-dir "$LLAMA_SERVER_MODELS_DIR" \
       --local-dir-use-symlinks False
   else
@@ -149,7 +149,7 @@ from huggingface_hub import hf_hub_download
 
 hf_hub_download(
     repo_id="TheBloke/TinyLlama-1.1B-Chat-v1.0-GGUF",
-    filename="TinyLlama-1.1B-Chat-v1.0.Q4_K_M.gguf",
+    filename="tinyllama-1.1b-chat-v1.0.Q4_K_M.gguf",
     local_dir=r"""$LLAMA_SERVER_MODELS_DIR""",
     local_dir_use_symlinks=False,
 )